Move templateLiterals visitors into a new subfolder
