+++ src/visitors/templateLiterals/index.js
@@ -1,17 +1,17 @@
 import {
   useCSSPreprocessor,
   useTranspileTemplateLiterals
-} from '../utils/options'
+} from '../../utils/options'
 
-import preprocessTemplateLiterals from './preprocessTemplateLiterals'
-import transpileTemplateLiterals from './transpileTemplateLiterals'
+import preprocess from './preprocess'
+import transpile from './transpile'
 
 export default (path, state) => {
   // We can only do one or the other, but preprocessing
   // disables the normal transpilation, obviously
   if (useCSSPreprocessor(state)) {
-    preprocessTemplateLiterals(path, state)
+    preprocess(path, state)
   } else if (useTranspileTemplateLiterals(state)) {
-    transpileTemplateLiterals(path, state)
+    transpile(path, state)
   }
 }
